Simplify time helper functions
Some of the logic around detecting if we're in the morning or afternoon
was a little convoluted and difficult to understand. This change reduces
the complexity of these functions.

Additionally, business hours on-call engineers are marked as AFK at the
end of the day rather than at midnight as was previously the case.

This change is necessary to try and identify where ailefroide is
breaking in adding/removing members from the lists

This change addresses the need by:

- Simplifying functions in the `calendar` set

As a result of this change, in local testing the following side effects
were observed:

- None